+++ deepfellow/suite/__init__.py
@@ -1,0 +1,18 @@
+# DeepFellow Software Framework.
+# Copyright © 2026 Simplito sp. z o.o.
+#
+# This file is part of the DeepFellow Software Framework (https://deepfellow.ai).
+# This software is Licensed under the DeepFellow Free License.
+#
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Collect suite commands."""
+
+import typer
+
+from .install import app as install_app
+
+app = typer.Typer()
+
+app.add_typer(install_app)